refactor: compartmentalize Subject Period code
